+++ indico/modules/events/abstracts/schemas.py
@@ -10,6 +10,7 @@ from marshmallow.fields import Float, Nested
 from indico.core.marshmallow import mm
 from indico.modules.events.abstracts.models.abstracts import Abstract
 from indico.modules.events.abstracts.models.comments import AbstractComment
+from indico.modules.events.abstracts.models.files import AbstractFile
 from indico.modules.events.abstracts.models.persons import AbstractPersonLink
 from indico.modules.events.abstracts.models.review_questions import AbstractReviewQuestion
 from indico.modules.events.abstracts.models.review_ratings import AbstractReviewRating
@@ -17,6 +18,8 @@ from indico.modules.events.abstracts.models.reviews import AbstractReview
 from indico.modules.events.contributions.schemas import ContributionFieldValueSchema, contribution_type_schema_basic
 from indico.modules.events.tracks.schemas import track_schema_basic
 from indico.modules.users.schemas import AffiliationSchema, BasicUserSchema
+from indico.util.mimetypes import icon_from_mimetype
+from indico.web.flask.util import url_for
 
 
 _basic_abstract_fields = ('id', 'friendly_id', 'title')
@@ -69,6 +72,15 @@ class AbstractPersonLinkSchema(mm.SQLAlchemyAutoSchema):
                   'address', 'phone', 'is_speaker', 'author_type')
 
 
+class AbstractFileSchema(mm.SQLAlchemyAutoSchema):
+    icon = mm.Function(lambda af: icon_from_mimetype(af.content_type))
+    download_url = mm.Function(lambda af: url_for('abstracts.download_attachment', af))
+
+    class Meta:
+        model = AbstractFile
+        fields = ('id', 'content_type', 'md5', 'filename', 'icon', 'download_url')
+
+
 class AbstractSchema(mm.SQLAlchemyAutoSchema):
     submitter = Nested(BasicUserSchema)
     judge = Nested(BasicUserSchema)
@@ -82,6 +94,7 @@ class AbstractSchema(mm.SQLAlchemyAutoSchema):
     reviewed_for_tracks = Nested(track_schema_basic, many=True)
     persons = Nested(AbstractPersonLinkSchema, attribute='person_links', many=True)
     custom_fields = Nested(ContributionFieldValueSchema, attribute='field_values', many=True)
+    files = Nested(AbstractFileSchema, many=True)
     score = Float()
     comments = Nested(AbstractCommentSchema, many=True)
     reviews = Nested(AbstractReviewSchema, many=True)
@@ -97,7 +110,7 @@ class AbstractSchema(mm.SQLAlchemyAutoSchema):
                   'submitted_contrib_type', 'accepted_contrib_type',
                   'accepted_track', 'submitted_for_tracks', 'reviewed_for_tracks',
                   'duplicate_of', 'merged_into',
-                  'persons', 'custom_fields',
+                  'persons', 'custom_fields', 'files',
                   'score', 'comments', 'reviews')